storage-element/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6  # For file uploads

# Configuration
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis (optional)
redis==5.0.1

# Storage
boto3==1.29.7  # For S3 storage
botocore==1.32.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2  # For FastAPI testing

# Security
pyjwt==2.8.0
cryptography==41.0.7

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
alembic==1.13.0
